init routing (landing, authorization)
