+++ src/app/app.component.html
@@ -1,41 +1,1 @@
-<app-header></app-header>
-<hr>
-<app-logo [logoParameters]="objLogoParameters.big"></app-logo>
-<app-logo [logoParameters]="objLogoParameters.small"></app-logo>
-<app-logo [logoParameters]="objLogoParameters.empty"></app-logo>
-<app-logo [logoParameters]="objLogoParameters.loader"></app-logo>
-<hr>
-<app-button tooltip="Hello" [buttonParameters]="objButtonParameters.big"></app-button>
-<app-button [buttonParameters]="objButtonParameters.medium"></app-button>
-<app-button [buttonParameters]="objButtonParameters.disable"></app-button>
-<app-button [buttonParameters]="objButtonParameters.small"></app-button>
-<app-button [buttonParameters]="objButtonParameters.smallEmpty"></app-button>
-<app-button [buttonParameters]="objButtonParameters.login"></app-button>
-<app-button [buttonParameters]="objButtonParameters.signUp"></app-button>
-<hr>
-<app-input [inputParameters]="objInputParameters.email"></app-input>
-<app-input [inputParameters]="objInputParameters.text"></app-input>
-<app-input [inputParameters]="objInputParameters.password"></app-input>
-<app-input [inputParameters]="objInputParameters.textarea"></app-input>
-<hr>
-<app-arrow-down></app-arrow-down>
-<hr>
-<app-button-toggle></app-button-toggle>
-<hr>
-<app-preview-card [previewCardParameters]="objPreviewCardParameters.developer"></app-preview-card>
-<app-preview-card [previewCardParameters]="objPreviewCardParameters.productOwner"></app-preview-card>
-<hr>
-<app-landing-guide [landingGuideParameters]="arrLandingGuideParameters"></app-landing-guide>
-<hr>
-<app-benefits-card [benefitsCardParameters]="objBenefitsCardParameters"></app-benefits-card>
-<hr>
-<app-select [selectParameters]="objSelectParameters"></app-select>
-<hr>
-<app-social-list [iconParameters]="arrSocialListParameters"></app-social-list>
-<hr>
-<app-role-toggle></app-role-toggle>
-<hr>
-<app-checkbox [checkboxParameters]="objCheckboxParameters"></app-checkbox>
-<hr>
-<app-project-card [projectCardParameters]="objProjectCardParameters.developer"></app-project-card>
-<app-project-card [projectCardParameters]="objProjectCardParameters.productOwner"></app-project-card>
+<router-outlet></router-outlet>
