+++ src/app/authorization/authorization-routing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { AuthorizationComponent } from './component/authorization.component';
+
+const routes: Routes = [
+  { path: '', component: AuthorizationComponent }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class AuthorizationRoutingModule {
+}
