+++ src/app/landing/component/landing.component.ts
