+++ src/app/landing/landing-routing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { LandingComponent } from './component/landing.component';
+
+const routes: Routes = [
+  { path: '', component: LandingComponent }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class LandingRoutingModule {
+}
