+++ src/app/landing/landing.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { LandingComponent } from './component/landing.component';
+import { LandingRoutingModule } from './landing-routing.module';
+
+@NgModule({
+  declarations: [
+    LandingComponent
+  ],
+  imports: [
+    CommonModule,
+    LandingRoutingModule
+  ],
+  exports: [
+    LandingComponent
+  ]
+})
+export class LandingModule { }
