--- src/app/shared/authorization/authorization.module.ts
@@ -1,15 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { AuthorizationComponent } from './component/authorization.component';
-
-
-
-@NgModule({
-  declarations: [
-    AuthorizationComponent
-  ],
-  imports: [
-    CommonModule
-  ]
-})
-export class AuthorizationModule { }
