--- src/app/shared/landing/landing.module.ts
@@ -1,15 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { LandingComponent } from './component/landing.component';
-
-
-
-@NgModule({
-  declarations: [
-    LandingComponent
-  ],
-  imports: [
-    CommonModule
-  ]
-})
-export class LandingModule { }
